Conflicts:
	mesalib/src/mesa/drivers/dri/swrast/swrast.c
	mesalib/src/mesa/main/.gitignore
	mesalib/src/mesa/main/queryobj.c

+++ fontconfig/src/fclist.c
@@ -270,7 +270,7 @@ FcListValueHash (FcValue    *value)
     case FcTypeCharSet:
 	return FcCharSetCount (v.u.c);
     case FcTypeFTFace:
-	return (long) v.u.f;
+	return (intptr_t) v.u.f;
     case FcTypeLangSet:
 	return FcLangSetHash (v.u.l);
     case FcTypeRange:
